+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * FixationIndexDependentModel
   * Bugfix: The CC of a constant saliency map wrt to a nonconstant one
     now returns zero (instead of nan as previously).
+  * Feature: Added keyword argument `attributes` to `Fixations` constructor
 * 0.2.19:
   * added pytorch implementation for optimization of similarity metric as alternative
     to tensorflow implementation which still uses tensorflow 1.x